Add files via upload
🌟 Excited to Share: My GuessNumber Program Developed at Encryptix! 🌟

During my internship at Encryptix, I created a fun and interactive GuessNumber program in Java. This project was an excellent opportunity to apply my coding skills and learn more about game development.

🎮 GuessNumber Features:

Prompts the user to guess a number within a specific range
Provides feedback if the guess is too high or too low
Continues until the user guesses the correct number
Tracks the number of attempts
🔹 Internship Role: Java Programming Intern
🔹 Company: Encryptix

I'm grateful for the opportunity to work with the amazing team at Encryptix. This experience has greatly enhanced my problem-solving skills and understanding of user interaction.

#Encryptix #JavaProgramming #SoftwareDevelopment #InternshipExperience #LearningAndGrowth #GuessNumber